Refactor application level modules to utilize an ActionStore
