+++ radixdlt-java/src/test/java/com/radixdlt/client/application/translate/TransferTokensActionTranslatorTest.java
@@ -4,37 +4,15 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.radixdlt.client.application.identity.RadixIdentity;
-import com.radixdlt.client.application.objects.TokenTransfer;
-import com.radixdlt.client.core.atoms.Atom;
-import io.reactivex.observers.TestObserver;
 import java.math.BigDecimal;
 import org.junit.Test;
 import com.radixdlt.client.application.actions.TransferTokensAction;
 import com.radixdlt.client.atommodel.tokens.TokenClassReference;
 import com.radixdlt.client.core.RadixUniverse;
 import com.radixdlt.client.atommodel.accounts.RadixAddress;
-import com.radixdlt.client.core.crypto.ECPublicKey;
 import java.util.Collections;
 
 public class TransferTokensActionTranslatorTest {
-	@Test
-	public void testSendToSelfTest() {
-		RadixUniverse universe = mock(RadixUniverse.class);
-		Atom atom = mock(Atom.class);
-		ECPublicKey myKey = mock(ECPublicKey.class);
-		RadixAddress myAddress = mock(RadixAddress.class);
-		when(universe.getAddressFrom(myKey)).thenReturn(myAddress);
-		TokenClassReference tokenClassReference = mock(TokenClassReference.class);
-		when(atom.tokenSummary()).thenReturn(Collections.singletonMap(tokenClassReference,
-			Collections.singletonMap(myKey, 0L)
-		));
-
-		TokenTransferTranslator tokenTransferTranslator = new TokenTransferTranslator(universe);
-		TestObserver<TokenTransfer> testObserver = TestObserver.create();
-		tokenTransferTranslator.fromAtom(atom, mock(RadixIdentity.class)).subscribe(testObserver);
-		testObserver.assertValue(transfer -> myAddress.equals(transfer.getFrom()) && myAddress.equals(transfer.getTo()));
-	}
 
 	@Test
 	public void createTransactionWithNoFunds() {
